+++ src/dropdown/list/createlistdropdown.js
@@ -30,10 +30,9 @@ import clickOutsideHandler from '../../bindings/clickoutsidehandler';
  *
  *		const dropdown = createListDropdown( model, locale );
  *
- *		dropdown.init();
- *
  *		// Will render a dropdown labeled "A dropdown" with a list in the panel
  *		// containing two items.
+ *		dropdown.render()
  *		document.body.appendChild( dropdown.element );
  *
  * The model instance remains in control of the dropdown after it has been created. E.g. changes to the
@@ -53,7 +52,6 @@ import clickOutsideHandler from '../../bindings/clickoutsidehandler';
  */
 export default function createListDropdown( model, locale ) {
 	const dropdownView = createDropdown( model, locale );
-
 	const listView = dropdownView.listView = new ListView( locale );
 
 	listView.items.bindTo( model.items ).using( itemModel => {