+++ src/editorui/boxed/boxededitoruiview.js
@@ -8,8 +8,8 @@
  */
 
 import EditorUIView from '../../editorui/editoruiview';
+import LabelView from '../../label/labelview';
 import uid from '@ckeditor/ckeditor5-utils/src/uid';
-import Template from '../../template';
 
 /**
  * The boxed editor UI view class. This class represents an editor interface
@@ -26,24 +26,7 @@ export default class BoxedEditorUIView extends EditorUIView {
 	constructor( locale ) {
 		super( locale );
 
-		const t = this.t;
 		const ariaLabelUid = uid();
-
-		/**
-		 * The UI's width.
-		 *
-		 * @observable
-		 * @member {Number} #width
-		 */
-		this.set( 'width', null );
-
-		/**
-		 * The UI's height.
-		 *
-		 * @observable
-		 * @member {Number} #height
-		 */
-		this.set( 'height', null );
 
 		/**
 		 * Collection of the child views located in the top (`.ck-editor__top`)
@@ -63,7 +46,16 @@ export default class BoxedEditorUIView extends EditorUIView {
 		 */
 		this.main = this.createCollection();
 
-		this.template = new Template( {
+		/**
+		 * Voice label of the UI.
+		 *
+		 * @protected
+		 * @readonly
+		 * @member {module:ui/view~View} #_voiceLabelView
+		 */
+		this._voiceLabelView = this._createVoiceLabel( ariaLabelUid );
+
+		this.setTemplate( {
 			tag: 'div',
 
 			attributes: {
@@ -74,22 +66,12 @@ export default class BoxedEditorUIView extends EditorUIView {
 				],
 				role: 'application',
 				dir: 'ltr',
-				lang: locale.lang,
+				lang: locale.language,
 				'aria-labelledby': `cke-editor__aria-label_${ ariaLabelUid }`
 			},
 
 			children: [
-				{
-					tag: 'span',
-					attributes: {
-						id: `cke-editor__aria-label_${ ariaLabelUid }`,
-						class: 'cke-voice-label'
-					},
-					children: [
-						// TODO: Editor name?
-						t( 'Rich Text Editor' )
-					]
-				},
+				this._voiceLabelView,
 				{
 					tag: 'div',
 					attributes: {
@@ -108,5 +90,27 @@ export default class BoxedEditorUIView extends EditorUIView {
 				}
 			]
 		} );
+	}
+
+	/**
+	 * Creates a voice label view instance.
+	 *
+	 * @private
+	 * @returns {module:ui/label/labelview~LabelView}
+	 */
+	_createVoiceLabel( ariaLabelUid ) {
+		const t = this.t;
+		const voiceLabel = new LabelView();
+
+		voiceLabel.text = t( 'Rich Text Editor' );
+
+		voiceLabel.extendTemplate( {
+			attributes: {
+				id: `cke-editor__aria-label_${ ariaLabelUid }`,
+				class: 'cke-voice-label'
+			}
+		} );
+
+		return voiceLabel;
 	}
 }