+++ tests/editorui/boxed/boxededitoruiview.js
@@ -12,20 +12,18 @@ describe( 'BoxedEditorUIView', () => {
 
 	beforeEach( () => {
 		view = new BoxedEditorUIView( new Locale( 'en' ) );
+		view.render();
 		element = view.element;
+	} );
 
-		return view.init();
+	afterEach( () => {
+		view.destroy();
 	} );
 
 	describe( 'constructor()', () => {
 		it( 'adds view collections', () => {
 			expect( view.top ).to.be.instanceof( ViewCollection );
 			expect( view.main ).to.be.instanceof( ViewCollection );
-		} );
-
-		it( 'sets "width" and "height" attributes', () => {
-			expect( view.width ).to.equal( null );
-			expect( view.height ).to.equal( null );
 		} );
 
 		it( 'bootstraps the view element from template', () => {