+++ tests/manual/icon/icon.js
@@ -7,8 +7,6 @@
 
 import IconView from '../../../src/icon/iconview';
 
-import '@ckeditor/ckeditor5-theme-lark/theme/theme.scss';
-
 const wrapper = document.querySelector( '#inline-svg' );
 
 const icon = `
@@ -59,8 +57,8 @@ addCase( iconWrapper );
 function renderIcon( content, size, color ) {
 	const iconView = new IconView();
 
+	iconView.render();
 	iconView.content = content;
-	iconView.init();
 
 	if ( size ) {
 		iconView.element.style.width = `${ size }px`;