+++ tests/toolbar/contextual/contextualtoolbar.js
@@ -10,11 +10,12 @@ import ToolbarView from '../../../src/toolbar/toolbarview';
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold';
 import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic';
+import Underline from '@ckeditor/ckeditor5-basic-styles/src/underline';
 import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
 
 import { setData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model.js';
 
-/* global document, setTimeout */
+/* global document, setTimeout, window */
 
 describe( 'ContextualToolbar', () => {
 	let sandbox, editor, contextualToolbar, balloon, editorElement;
@@ -43,8 +44,10 @@ describe( 'ContextualToolbar', () => {
 
 				// Focus the engine.
 				editor.editing.view.isFocused = true;
+				editor.editing.view.getDomRoot().focus();
 
-				contextualToolbar.toolbarView.init();
+				// Remove all selection ranges from DOM before testing.
+				window.getSelection().removeAllRanges();
 			} );
 	} );
 
@@ -69,6 +72,28 @@ describe( 'ContextualToolbar', () => {
 
 	it( 'should create components from config', () => {
 		expect( contextualToolbar.toolbarView.items ).to.length( 2 );
+	} );
+
+	it( 'should accept the extended format of the toolbar config', () => {
+		const editorElement = document.createElement( 'div' );
+		document.body.appendChild( editorElement );
+
+		return ClassicTestEditor
+			.create( editorElement, {
+				plugins: [ Paragraph, Bold, Italic, Underline, ContextualToolbar ],
+				contextualToolbar: {
+					items: [ 'bold', 'italic', 'underline' ]
+				}
+			} )
+			.then( editor => {
+				const contextualToolbar = editor.plugins.get( ContextualToolbar );
+
+				expect( contextualToolbar.toolbarView.items ).to.length( 3 );
+
+				editorElement.remove();
+
+				return editor.destroy();
+			} );
 	} );
 
 	it( 'should fire internal `_selectionChangeDebounced` event 200 ms after last selection change', done => {
@@ -98,12 +123,12 @@ describe( 'ContextualToolbar', () => {
 				// Still not yet.
 				sinon.assert.notCalled( spy );
 
-				// Another 101 ms waiting.
+				// Another waiting.
 				setTimeout( () => {
 					// And here it is.
 					sinon.assert.calledOnce( spy );
 					done();
-				}, 100 );
+				}, 110 );
 			}, 101 );
 		}, 100 );
 	} );
@@ -263,7 +288,7 @@ describe( 'ContextualToolbar', () => {
 			let showSpy;
 
 			beforeEach( () => {
-				showSpy = sinon.spy( contextualToolbar, 'show' );
+				showSpy = sandbox.spy( contextualToolbar, 'show' );
 			} );
 
 			it( 'should not be called when the editor is not focused', () => {
@@ -416,7 +441,7 @@ describe( 'ContextualToolbar', () => {
 
 			contextualToolbar.fire( '_selectionChangeDebounced' );
 
-			const stub = sinon.stub( balloon, 'visibleView' ).get( () => contextualToolbar.toolbarView );
+			const stub = sandbox.stub( balloon, 'visibleView' ).get( () => contextualToolbar.toolbarView );
 
 			sinon.assert.calledOnce( showPanelSpy );
 			sinon.assert.notCalled( hidePanelSpy );
@@ -434,7 +459,7 @@ describe( 'ContextualToolbar', () => {
 
 			contextualToolbar.fire( '_selectionChangeDebounced' );
 
-			const stub = sinon.stub( balloon, 'visibleView' ).get( () => null );
+			const stub = sandbox.stub( balloon, 'visibleView' ).get( () => null );
 
 			sinon.assert.calledOnce( showPanelSpy );
 			sinon.assert.notCalled( hidePanelSpy );
@@ -450,7 +475,7 @@ describe( 'ContextualToolbar', () => {
 
 	describe( 'show event', () => {
 		it( 'should fire `show` event just before panel shows', () => {
-			const spy = sinon.spy();
+			const spy = sandbox.spy();
 
 			contextualToolbar.on( 'show', spy );
 			setData( editor.document, '<paragraph>b[a]r</paragraph>' );